+++ lang/attribute/program/src/declare_program/mod.rs
@@ -11,7 +11,7 @@ use syn::parse::{Parse, ParseStream};
 use common::gen_docs;
 use mods::{
     accounts::gen_accounts_mod, client::gen_client_mod, constants::gen_constants_mod,
-    cpi::gen_cpi_mod, errors::gen_errors_mod, events::gen_events_mod, internal::gen_internal_mod,
+    cpi::gen_cpi_mod, error::gen_error_mod, events::gen_events_mod, internal::gen_internal_mod,
     parsers::gen_parsers_mod, program::gen_program_mod, types::gen_types_mod,
 };
 
@@ -61,7 +61,7 @@ fn gen_program(idl: &Idl, name: &syn::Ident) -> proc_macro2::TokenStream {
     let accounts_mod = gen_accounts_mod(idl);
     let events_mod = gen_events_mod(idl);
     let types_mod = gen_types_mod(idl);
-    let errors_mod = gen_errors_mod(idl);
+    let error_mod = gen_error_mod(idl);
 
     // Clients
     let cpi_mod = gen_cpi_mod(idl);
@@ -91,7 +91,7 @@ fn gen_program(idl: &Idl, name: &syn::Ident) -> proc_macro2::TokenStream {
             #accounts_mod
             #events_mod
             #types_mod
-            #errors_mod
+            #error_mod
 
             #cpi_mod
             #client_mod