+++ lang/attribute/program/src/declare_program/mods/mod.rs
@@ -2,7 +2,7 @@ pub mod accounts;
 pub mod client;
 pub mod constants;
 pub mod cpi;
-pub mod errors;
+pub mod error;
 pub mod events;
 pub mod internal;
 pub mod parsers;